部门删除会先检查其中是否还有用户

+++ src/main/java/com/fishpound/accountservice/controller/DepartmentController.java
@@ -36,6 +36,10 @@ public class DepartmentController {
 
     @DeleteMapping("/{id}")
     public JsonResult deleteDepartment(@PathVariable(value = "id") String id){
+        Department department = departmentService.findById(id);
+        if(department.getUserInfoSet() != null && department.getUserInfoSet().size() > 0){
+            return ResultTool.fail("要删除的部门中还存在用户，请先将用户转移到其他部门");
+        }
         departmentService.deleteDept(id);
         return ResultTool.success();
     }
